moved AuthorizationUI.kt to its designated package
